refactor(api): streamline imports by consolidating AttachmentSource type in sendMessage and uploadAttachment files
